+++ client/react-utils/components/modal-container/component.jsx
@@ -9,14 +9,25 @@ const Component = (props) => {
         : null
     ;
 
+    const header = title
+        ? <Modal.Header closeButton>{title}</Modal.Header>
+        : null
+    ;
+
+    const footer = props.footerButtons && props.footerButtons.length
+        ? <Modal.Footer>TODO: Modal.Footer</Modal.Footer>
+        : null
+    ;
+
     return (
-        <Modal bsSize="large" className={props.className} show={props.show} onHide={props.onHide}>
-            <Modal.Header closeButton>
-                {title}
-            </Modal.Header>
+        <Modal bsSize="large" className={`${props.className} warpjs-react-modal`} show={props.show} onHide={props.onHide}
+            data-warpjs-modal={props.id}
+            >
+            {header}
             <Modal.Body>
                 {props.children}
             </Modal.Body>
+            {footer}
         </Modal>
     );
 };
@@ -25,10 +36,12 @@ Component.displayName = 'ModalContainer';
 
 Component.propTypes = {
     children: PropTypes.element,
+    className: PropTypes.string,
+    footerButtons: PropTypes.array,
+    id: PropTypes.string.isRequired,
     onHide: PropTypes.func,
     show: PropTypes.bool,
     title: PropTypes.string,
-    className: PropTypes.string
 };
 
 export default errorBoundary(Component);